peer resolution refactor
